fix(groomer): model interfaces 폴더 조정 및 견적서 라우팅 버튼 추가
