+++ apps/groomer/src/pages/estimate/index.styles.ts
@@ -2,6 +2,7 @@ import { theme } from '@daengle/design-system';
 import { css } from '@emotion/react';
 
 export const wrapper = css`
+  height: 100vh;
   padding-bottom: 104px;
 
   background-color: ${theme.colors.background};
@@ -16,5 +17,7 @@ export const listContainer = css`
   flex: 1;
   overflow-y: auto;
 
-  padding: 18px 18px 12px 0;
+  padding: 18px 18px 104px 0;
+
+  background-color: ${theme.colors.background};
 `;
